Refactor config to process arguments by dispatchers
Processing of all the arguments of `.option` and `.param`
definitions was extracted to separate dispatchers.

Now we have more control of the processing and can
provide additional rules by adding new dispatchers.
This can be used for:
- fixing longstanding issue about too strict
  arity check for dry-types
- adding support for 'array types' like:
  ```ruby
  option :foo, [method(:String)]
  ```
  as an alternative to
  ```ruby
  option :foo, ->(v) { Array(v).map { |item| String(item) } }
  ```
- adding support for nested struct definitions like:
  ```ruby
  option :foo do
    option :bar
    option :baz
  end
  ```

The downcase of this change is that it adds more
instantiations of objects (several dispatchers for
every definition). But because this job is done in
loading time I count this as not important.