Support nested type definitions
From now and on you can define a type/coercer
of the argument values via the block:

```ruby
class Test
  extend Dry::Initializer

  param :foo do
    option :bar do
      option :baz, proc(&:to_s)
    end
  end
end

test = Test.new(bar: { baz: :qux }
test.foo.bar.baz # => "qux"
```

Inside the block only `option`, not `param` helper is supported.
All the other settings works in exactly the same way as for
the upper level.

The only restriction is that you cannot define a coercer
in both the block and the argument. Both the following defs are invalid:

```ruby
class Test
  extend Dry::Initializer

  param :foo, proc(&:to_s) do
    option :qux
  end

  param :foo, type: proc(&:to_s) do
    option :qux
  end
end
```

This is because we don't know what coercer to be used.

The only EXCLUSION from this rule concerns wrapped types.
The following definition is correct:

```ruby
class Test
  extend Dry::Initializer

  param :foo, [] do
    option :bar, proc(&:to_s)
  end
end

test = Test.new(bar: 42)
test.foo # => ["42"]
```